ported over code from dev componenets to Cart and CartItems
